# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GasPrice {
  price: number;
  state: string;
  source: string;
}

export interface VehicleYear {
  value: string;
  text: string;
}

export type VehicleYearList = VehicleYear[];

export interface VehicleMake {
  value: string;
  text: string;
}

export type VehicleMakeList = VehicleMake[];

export interface VehicleModel {
  value: string;
  text: string;
}

export type VehicleModelList = VehicleModel[];

export interface VehicleTrim {
  id: string;
  text: string;
}

export type VehicleTrimList = VehicleTrim[];

export interface VehicleMpg {
  combined: number;
  city: number;
  highway: number;
  fuelType: string;
}

export interface TripDistance {
  miles: number;
  duration: string;
}

export interface ApiError {
  error: string;
}

export type GetGasPriceParams = {
zip: string;
};

export type GetVehicleMakesParams = {
year: string;
};

export type GetVehicleModelsParams = {
year: string;
make: string;
};

export type GetVehicleTrimsParams = {
year: string;
make: string;
model: string;
};

export type GetVehicleMpgParams = {
id: string;
};

export type GetTripDistanceParams = {
origin: string;
destination: string;
};