# lib/api-zod/src/generated/types/gasPrice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GasPrice {
  price: number;
  state: string;
  source: string;
}