# lib/api-zod/src/generated/types/vehicleYearList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VehicleYear } from './vehicleYear';

export type VehicleYearList = VehicleYear[];